PMS/requirements.txt
asyncio
paho-mqtt>=1.6.1
APScheduler>=3.10.4
pymodbus>=3.5.2
PyYAML>=6.0.1 